% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mewto.R
\name{mewtoApp}
\alias{mewtoApp}
\title{Model Evaluation with Weighted Threshold Optimization App}
\usage{
mewtoApp(actuals, probabilities)
}
\arguments{
\item{actuals}{Data of factor type with two levels: "yes" for positive and
"no" for negative.}

\item{probabilities}{Data of numeric type which should represent the
probabilities of realization of the positive category.}
}
\description{
This function launches a Shiny application where the user can interactively
manipulate the threshold used in binary classification and view the
associated metrics, confusion matrix and ROC curve. The app also allows
for optimal threshold calculation according to a weighted version of
Youden's J-statistic.
}
\details{
In the calculation of the optimal threshold, a weighted version of
Youden's J-statistic (Youden, 1950) is employed. The optimal cut-off is the
threshold that maximizes the distance to the identity (diagonal) line. The
function maximizes the metric 2 * (w * sensitivity + (1 - w) * specificity),
where w is the "weight" parameter chosen through the second slider. After
setting the desired value on the optimal threshold slider, the user must
press the button "Set" to update the threshold slider above with the optimal
value.
}
\examples{
mewtoApp(actuals, probabilities)
}
\author{
Alexandru Monahov, \email{alexandrudezv@gmail.com}
}
\keyword{confusion,}
\keyword{evaluation,}
\keyword{matrix,}
\keyword{optimum,}
\keyword{roc}
\keyword{threshold,}
