% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mewto.R
\name{mewtoThresh}
\alias{mewtoThresh}
\title{Optimal threshold calculation using weighted Youden's J-statistic}
\usage{
mewtoThresh(actuals, probabilities, weight = 0.5)
}
\arguments{
\item{actuals}{Data of factor type with two levels: "yes" for positive and
"no" for negative.}

\item{probabilities}{Data of numeric type which should represent the
probabilities of realization of the positive category.}

\item{weight}{A numeric value corresponding to the importance attributed to sensitivity, or formulated
differently, to the maximization of the true positives rate. This value should be in the interval \link{0 ; 1}.}
}
\description{
This function calculates the optimal threshold according to a weighted
version of Youden's J-statistic.
}
\details{
In the calculation of the optimal threshold, a weighted version of
Youden's J-statistic (Youden, 1950) is employed. The optimal cut-off is the
threshold that maximizes the distance to the identity (diagonal) line. The
function maximizes the metric (w * sensitivity + (1 - w) * specificity),
where w is the "weight" parameter. Metrics will automatically be recalculated
based on the user's selection.
}
\examples{
mewtoThresh(actuals, probabilities, weight = 0.5) # Original Youden's J-statistic.
mewtoThresh(actuals, probabilities, weight = 0.9) # Favor sensitivity greatly over specificity.
}
\author{
Alexandru Monahov, \email{alexandrudezv@gmail.com}
}
\keyword{evaluation,}
\keyword{optimum,}
\keyword{threshold,}
\keyword{weighted}
\keyword{youden,}
